refactor(pty): move tests into dedicated directory
