+++ station/package.json
@@ -20,9 +20,9 @@
     "host:dev": "bun run link:station && bun run repair:node-pty && bun src/host/e2eReattachDemo.ts --dev --hold",
     "host:list": "bun run link:station && bun src/host/listHostAgents.ts",
     "test:agents": "bun run repair:node-pty && STATION_PTY_SMOKE=1 STATION_PTY_SMOKE_TUI=1 bun test src/input/stationInput.e2e.test.tsx src/terminal/ptyPipeline.smoke.test.ts",
-    "test:pty": "bun run repair:node-pty && STATION_PTY_SMOKE=1 bun test src/terminal/pty/childPtyEnvironment.test.ts src/terminal/pty/localPtyTerminal.test.ts src/terminal/pty/auxShellPlacement.test.ts src/terminal/ptyPipeline.smoke.test.ts src/host/ptyTable.smoke.test.ts src/terminal/pty/hostReplaySizing.test.ts src/terminal/pty/localPtyBridgeHardening.test.ts src/terminal/pty/localPtyBridgeOrphan.test.ts src/host/ptyTableRegistry.test.ts src/host/orphanBridges.test.ts src/host/orphanAdoption.smoke.test.ts src/host/hostHandoff.smoke.test.ts src/host/hostHandoff.packaging.smoke.test.ts src/input/stationInput.e2e.test.tsx",
+    "test:pty": "bun run repair:node-pty && STATION_PTY_SMOKE=1 bun test src/terminal/pty/test/childPtyEnvironment.test.ts src/terminal/pty/test/localPtyTerminal.test.ts src/terminal/pty/test/auxShellPlacement.test.ts src/terminal/ptyPipeline.smoke.test.ts src/host/ptyTable.smoke.test.ts src/terminal/pty/test/hostReplaySizing.test.ts src/terminal/pty/test/localPtyBridgeHardening.test.ts src/terminal/pty/test/localPtyBridgeOrphan.test.ts src/host/ptyTableRegistry.test.ts src/host/orphanBridges.test.ts src/host/orphanAdoption.smoke.test.ts src/host/hostHandoff.smoke.test.ts src/host/hostHandoff.packaging.smoke.test.ts src/input/stationInput.e2e.test.tsx",
     "test:host:upgrade": "bun run repair:node-pty && STATION_PTY_SMOKE=1 bun test src/host/hostHandoff.smoke.test.ts src/host/hostHandoff.packaging.smoke.test.ts",
-    "test:pty:bun": "bun run build:ctty-helper && STATION_PTY_SMOKE=1 STATION_PTY_IMPL=bun bun test src/terminal/pty/terminalProcessEmitter.test.ts src/terminal/pty/childPtyEnvironment.test.ts src/terminal/pty/localPtyTerminal.test.ts src/terminal/pty/auxShellPlacement.test.ts src/terminal/pty/bunTerminalProcess.test.ts src/host/ptyTable.smoke.test.ts src/terminal/pty/hostReplaySizing.test.ts src/host/ptyTableRegistry.test.ts src/host/orphanBridges.test.ts",
+    "test:pty:bun": "bun run build:ctty-helper && STATION_PTY_SMOKE=1 STATION_PTY_IMPL=bun bun test src/terminal/pty/test/terminalProcessEmitter.test.ts src/terminal/pty/test/childPtyEnvironment.test.ts src/terminal/pty/test/localPtyTerminal.test.ts src/terminal/pty/test/auxShellPlacement.test.ts src/terminal/pty/test/bunTerminalProcess.test.ts src/host/ptyTable.smoke.test.ts src/terminal/pty/test/hostReplaySizing.test.ts src/host/ptyTableRegistry.test.ts src/host/orphanBridges.test.ts",
     "test:stress": "STATION_VT_STRESS=1 bun test src/terminal/vt/screen.stress.test.ts",
     "test:vt": "bun test src/terminal",
     "typecheck": "bun run link:station && tsc -p tsconfig.json --noEmit"
